+++ packages/projection-typeorm/test/operators/storeNftMetadata.test.ts
@@ -31,7 +31,8 @@ import {
   createRollForwardEventBasedOn,
   createStubBlockHeader,
   createStubProjectionSource,
-  createStubRollForwardEvent
+  createStubRollForwardEvent,
+  filterAssets
 } from './util';
 import { dummyLogger } from 'ts-log';
 import omit from 'lodash/omit';
@@ -257,7 +258,7 @@ describe('storeNftMetadata', () => {
       );
       expect(omit(storedNftMetadata, ['userTokenAsset', 'userTokenAssetId', 'id'])).toEqual({
         description: someNftMetadata.description || null,
-        files: someNftMetadata.files || null,
+        files: someNftMetadata.files,
         image: someNftMetadata.image,
         mediaType: someNftMetadata.mediaType || null,
         name: someNftMetadata.name,
@@ -321,6 +322,31 @@ describe('storeNftMetadata', () => {
       }),
       NftMetadataType.CIP25
     );
+
+    it('does not throw when name has null characters', async () => {
+      const events = chainSyncData(ChainSyncDataSet.AssetNameUtf8Problem);
+      await expect(
+        lastValueFrom(
+          project$(
+            filterAssets(events, [Cardano.AssetId('00740069006e0079002000640069006e006f0073002000230035003600350032')])
+          )
+        )
+        // throws 'invalid byte sequence for encoding "UTF8": 0x00'
+        // when asset name is not sanitized
+      ).resolves.not.toThrow();
+    });
+
+    it('does not throw when some field in otherProperties has null characters', async () => {
+      const events = chainSyncData(ChainSyncDataSet.ExtraDataNullCharactersProblem);
+      const assetId = Cardano.AssetId('65bcf672806de8a2335576339e801d41f3275c0c07dd6aadf2ea41d9000000000042414444');
+
+      // throws 'unsupported Unicode escape sequence'
+      // when otherProperties is not sanitized
+      await lastValueFrom(project$(filterAssets(events, [assetId])));
+
+      const metadata = await nftMetadataRepo.findOneBy({ userTokenAssetId: assetId });
+      expect(metadata!.otherProperties!.size).toBeGreaterThan(0);
+    });
   });
 
   describe('cip68', () => {
@@ -442,11 +468,24 @@ describe('storeNftMetadata', () => {
     expect(events.length).toBeGreaterThan(1);
     expect(await nftMetadataRepo.count()).toBeGreaterThan(0);
     const nftAssets = await nftMetadataRepo.find({ select: ['files'] });
-    const file = nftAssets.find((asset) => asset?.files!.length > 0)!.files![0];
+    const file = nftAssets.find((asset) => asset.files && asset.files.length > 0)!.files![0];
     expect(typeof file.mediaType).toBe('string');
     expect(typeof file.src).toBe('string');
     expect(['object', 'undefined'].includes(typeof file.otherProperties)).toBe(true);
     expect(['string', 'undefined'].includes(typeof file.name)).toBe(true);
+  });
+
+  it('projects metadata with missing "extra" field', async () => {
+    const events = chainSyncData(ChainSyncDataSet.MissingExtraDatumMetadataProblem);
+    const userTokenAssetId = Cardano.AssetId(
+      'e51fbae37cc032eab73861f52ccfa3291e1f4746b7a471628ae27012000de1406e6d6b724e4654386d6179'
+    );
+    const referenceNftAssetId = Cardano.AssetId(
+      'e51fbae37cc032eab73861f52ccfa3291e1f4746b7a471628ae27012000643b06e6d6b724e4654386d6179'
+    );
+    await lastValueFrom(project$(filterAssets(events, [referenceNftAssetId, userTokenAssetId])));
+    const metadata = await nftMetadataRepo.findOneBy({ userTokenAssetId });
+    expect(metadata).toBeTruthy();
   });
 });