+++ packages/wallet/src/Wallets/BaseWallet.ts
@@ -93,6 +93,7 @@ import { Cip30DataSignature } from '@cardano-sdk/dapp-connector';
 import { Ed25519PublicKey, Ed25519PublicKeyHex } from '@cardano-sdk/crypto';
 import {
   GenericTxBuilder,
+  GreedyTxEvaluator,
   InitializeTxProps,
   InitializeTxResult,
   InvalidConfigurationError,
@@ -552,6 +553,7 @@ export class BaseWallet implements ObservableWallet {
       : throwError(() => new InvalidConfigurationError('BaseWallet is missing a "handleProvider"'));
 
     this.util = createWalletUtil({
+      chainHistoryProvider: this.chainHistoryProvider,
       protocolParameters$: this.protocolParameters$,
       transactions: this.transactions,
       utxo: this.utxo
@@ -774,6 +776,7 @@ export class BaseWallet implements ObservableWallet {
         tip: () => this.#firstValueFromSettled(this.tip$),
         utxoAvailable: () => this.#firstValueFromSettled(this.utxo.available$)
       },
+      txEvaluator: new GreedyTxEvaluator(() => this.#firstValueFromSettled(this.protocolParameters$)),
       witnesser: this.witnesser
     };
   }