+++ packages/governance/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.5](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.10.4...@cardano-sdk/governance@0.10.5) (2024-05-20)
+
+**Note:** Version bump only for package @cardano-sdk/governance
+
 ## [0.10.4](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/governance@0.10.3...@cardano-sdk/governance@0.10.4) (2024-05-02)
 
 **Note:** Version bump only for package @cardano-sdk/governance