+++ packages/input-selection/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.31](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.30...@cardano-sdk/input-selection@0.12.31) (2024-05-20)
+
+**Note:** Version bump only for package @cardano-sdk/input-selection
+
 ## [0.12.30](https://github.com/input-output-hk/cardano-js-sdk/compare/@cardano-sdk/input-selection@0.12.29...@cardano-sdk/input-selection@0.12.30) (2024-05-02)
 
 **Note:** Version bump only for package @cardano-sdk/input-selection